form data

+++ database/object/minio.go
@@ -28,8 +28,8 @@ func NewStorage(ctx context.Context, mc *minio.Client, bucket, host string) (*Mi
 	}, nil
 }
 
-func (s *Minio) Add(ctx libctx.Context, fileName string, payload io.Reader, payloadLength int) (string, error) {
-	_, err := s.mc.PutObject(ctx, s.bucket, fileName, payload, int64(payloadLength), minio.PutObjectOptions{})
+func (s *Minio) Add(ctx libctx.Context, fileName string, payload io.Reader, payloadLength int64) (string, error) {
+	_, err := s.mc.PutObject(ctx, s.bucket, fileName, payload, payloadLength, minio.PutObjectOptions{})
 	if err != nil {
 		return "", fmt.Errorf("could not put object %s: %w", fileName, err)
 	}
