+++ service/image/interfaces.go
@@ -6,5 +6,5 @@ import (
 )
 
 type Storage interface {
-	Add(ctx libctx.Context, fileName string, payload io.Reader, payloadLength int) (string, error)
+	Add(ctx libctx.Context, fileName string, payload io.Reader, payloadLength int64) (string, error)
 }
